+++ plugins/builtin/source/content/tools/euclidean_alg.cpp
@@ -54,8 +54,8 @@ namespace hex::plugin::builtin {
 
                     overflow = true;
                 } else {
-                    gcdResult       = std::gcd<i128, i128>(a, b);
-                    lcmResult       = std::lcm<i128, i128>(a, b);
+                    gcdResult       = std::gcd<i64, i64>(a, b);
+                    lcmResult       = std::lcm<i64, i64>(a, b);
                     std::tie(p, q)  = extendedGcd(a, b);
 
                     overflow = false;
